refactor(article): set base package of application to "nettee.blolet.article.application"
