--- services/article/application/src/main/java/nettee/article/port/ArticleLikesCommandRepositoryPort.java
@@ -1,14 +1,0 @@
-package nettee.article.port;
-
-import nettee.blolet.article.domain.ArticleLikes;
-
-import java.util.Optional;
-
-public interface ArticleLikesCommandRepositoryPort {
-
-    Optional<ArticleLikes> findByProfileIdAndArticleId(String profileId, String articleId);
-
-    ArticleLikes save(ArticleLikes articleLikes);
-
-    void deleteByProfileIdAndArticleId(String profileId, String articleId);
-}
