--- services/article/application/src/main/java/nettee/article/port/ArticleQueryRepositoryPort.java
@@ -1,15 +1,0 @@
-package nettee.article.port;
-
-import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleDetail;
-import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleSummary;
-import org.springframework.data.domain.Slice;
-
-import java.time.Instant;
-import java.util.Optional;
-
-public interface ArticleQueryRepositoryPort {
-
-    Optional<ArticleDetail> findByArticleId(String articleId);
-
-    Slice<ArticleSummary> findAllByBlogId(String blogId, Instant lastCreatedAt, int size);
-}
