+++ services/article/application/src/main/java/nettee/blolet/article/application/port/ArticleLikesCommandRepositoryPort.java
@@ -1,0 +1,14 @@
+package nettee.blolet.article.application.port;
+
+import nettee.blolet.article.domain.ArticleLikes;
+
+import java.util.Optional;
+
+public interface ArticleLikesCommandRepositoryPort {
+
+    Optional<ArticleLikes> findByProfileIdAndArticleId(String profileId, String articleId);
+
+    ArticleLikes save(ArticleLikes articleLikes);
+
+    void deleteByProfileIdAndArticleId(String profileId, String articleId);
+}
