+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/ArticleLikesDeleteUseCase.java
@@ -1,4 +1,4 @@
-package nettee.article.usecase;
+package nettee.blolet.article.application.usecase;
 
 public interface ArticleLikesDeleteUseCase {
