+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/ArticleReadUseCase.java
@@ -1,0 +1,14 @@
+package nettee.blolet.article.application.usecase;
+
+import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleDetail;
+import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleSummary;
+import org.springframework.data.domain.Slice;
+
+import java.time.Instant;
+
+public interface ArticleReadUseCase {
+
+    ArticleDetail getArticle(String articleId);
+
+    Slice<ArticleSummary> getArticleList(String blogId, Instant lastCreatedAt, int size);
+}
