+++ services/article/driven/rdb/src/main/java/nettee/article/driven/rdb/persistence/ArticleQueryAdapter.java
@@ -2,7 +2,7 @@ package nettee.article.driven.rdb.persistence;
 
 import lombok.RequiredArgsConstructor;
 import nettee.article.driven.rdb.persistence.mapper.ArticleEntityMapper;
-import nettee.article.port.ArticleQueryRepositoryPort;
+import nettee.blolet.article.application.port.ArticleQueryRepositoryPort;
 import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleDetail;
 import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleSummary;
 import org.springframework.data.domain.PageRequest;
