+++ services/article/driving/web-mvc/src/main/java/nettee/article/driving/web/ArticleCommandApi.java
@@ -9,11 +9,11 @@ import nettee.article.driving.web.dto.ArticleCommandDto.ArticleCommandResponse;
 import nettee.article.driving.web.dto.ArticleCommandDto.ArticleCreateCommand;
 import nettee.article.driving.web.dto.ArticleLikesCommandDto.ArticleLikesCommandResponse;
 import nettee.article.driving.web.dto.ArticleLikesCommandDto.ArticleLikesCreateCommand;
-import nettee.article.usecase.ArticleCreateUseCase;
-import nettee.article.usecase.ArticleDeleteUseCase;
-import nettee.article.usecase.ArticleLikesDeleteUseCase;
-import nettee.article.usecase.ArticleLikesUpdateUseCase;
-import nettee.article.usecase.ArticleUpdateUseCase;
+import nettee.blolet.article.application.usecase.ArticleCreateUseCase;
+import nettee.blolet.article.application.usecase.ArticleDeleteUseCase;
+import nettee.blolet.article.application.usecase.ArticleLikesDeleteUseCase;
+import nettee.blolet.article.application.usecase.ArticleLikesUpdateUseCase;
+import nettee.blolet.article.application.usecase.ArticleUpdateUseCase;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
